refactor(ts): AudioPlayer
