--- source/components/AudioPlayer/index.js
@@ -1,3 +1,0 @@
-import './style/index.less';
-
-export default from './AudioPlayer';
