+++ source/components/AudioPlayer/index.tsx
@@ -1,0 +1,4 @@
+import AudioPlayer from './AudioPlayer';
+import './style/index.less';
+
+export default AudioPlayer;
